refactor(storage): split composite.rs into submodules (#179)
Break the monolithic composite.rs (2400+ lines) into focused submodules
to reduce its high betweenness centrality and improve maintainability:

- composite/mod.rs    — struct, constructor, accessors, StorageProvider impl
- composite/config.rs — RaidMode and CompositeConfig types
- composite/erasure.rs — Reed-Solomon encode/decode and shard I/O
- composite/fanout.rs  — fan_out, fan_out_void, try_first helpers
- composite/health.rs  — health tracking, probing, redundancy warnings
- composite/tests.rs   — all existing tests (unchanged)

Public API is identical; this is a pure structural refactor.
